fix identify on the client for hubspot
