+++ packages/client/components/Action/Action.tsx
@@ -1,7 +1,6 @@
 import {css, Global} from '@emotion/core'
 import React, {lazy, memo, Suspense} from 'react'
 import {Route, Switch} from 'react-router'
-import useCacheEmail from '~/hooks/useCacheEmail'
 import useServiceWorkerUpdater from '../../hooks/useServiceWorkerUpdater'
 import useTrebuchetEvents from '../../hooks/useTrebuchetEvents'
 import globalStyles from '../../styles/theme/globalStyles'
@@ -32,7 +31,6 @@ const InvitationLink = lazy(() =>
 )
 
 const Action = memo(() => {
-  useCacheEmail()
   useTrebuchetEvents()
   useServiceWorkerUpdater()
   return (
