+++ packages/client/components/AnalyticsPage.tsx
@@ -1,10 +1,23 @@
 /// <reference types="@types/segment-analytics" />
 
 import * as Sentry from '@sentry/browser'
+import graphql from 'babel-plugin-relay/macro'
 import {useEffect, useRef} from 'react'
+import {fetchQuery} from 'react-relay'
+import useAtmosphere from '~/hooks/useAtmosphere'
 import {LocalStorageKey} from '~/types/constEnums'
+import safeIdentify from '~/utils/safeIdentify'
+import {AnalyticsPageQuery} from '~/__generated__/AnalyticsPageQuery.graphql'
 import useScript from '../hooks/useScript'
 import makeHref from '../utils/makeHref'
+
+const query = graphql`
+  query AnalyticsPageQuery {
+    viewer {
+      email
+    }
+  }
+`
 
 declare global {
   interface Window {
@@ -13,14 +26,6 @@ declare global {
 }
 
 const dsn = window.__ACTION__.sentry
-
-let email: string | null = null
-const getEmail = () => {
-  if (!email) {
-    email = window.localStorage.getItem(LocalStorageKey.EMAIL)
-  }
-  return email
-}
 
 if (dsn) {
   Sentry.init({
@@ -31,6 +36,11 @@ if (dsn) {
   })
 }
 
+// page titles are changed in child components via useDocumentTitle, which fires after this
+// we must guarantee that this runs after useDocumentTitle
+// we can't move this into useDocumentTitle since the pathname may change without chaging the title
+const TIME_TO_RENDER_TREE = 100
+
 const AnalyticsPage = () => {
   const key = window.__ACTION__.segment
   if (!key) return null // development use
@@ -38,23 +48,43 @@ const AnalyticsPage = () => {
   const {href, pathname} = location
   const pathnameRef = useRef(pathname)
   const [isSegmentLoaded] = useScript(`https://cdn.segment.com/analytics.js/v1/${key}/analytics.min.js`, {crossOrigin: true})
+  const atmosphere = useAtmosphere()
+
+  useEffect(() => {
+    if (!isSegmentLoaded || !window.analytics) return
+    const token = window.localStorage.getItem(LocalStorageKey.APP_TOKEN_KEY)
+    // no token means authentication is required & authentication handles identify on its own
+    if (!token) return
+    const email = window.localStorage.getItem(LocalStorageKey.EMAIL)
+    if (email) {
+      safeIdentify(atmosphere.viewerId, email)
+      return
+    }
+    const cacheEmail = async () => {
+      const res = await fetchQuery<AnalyticsPageQuery>(atmosphere, query, {})
+      const nextEmail = res?.viewer?.email
+      if (!nextEmail) return
+      window.localStorage.setItem(LocalStorageKey.EMAIL, nextEmail)
+      safeIdentify(atmosphere.viewerId, nextEmail)
+    }
+    cacheEmail().catch()
+  }, [isSegmentLoaded])
+
   useEffect(() => {
     if (!isSegmentLoaded || !window.analytics) return
     const prevPathname = pathnameRef.current
     pathnameRef.current = pathname
-    // helmet sets titles async, so we have to wait awhile until it updates
     setTimeout(() => {
       const title = document.title || ''
       // This is the magic. Ignore everything after hitting the pipe
       const [pageName] = title.split(' | ')
       window.analytics.page(pageName, {
-        email: getEmail(),
         referrer: makeHref(prevPathname),
         title,
         path: pathname,
         url: href
       })
-    }, 300)
+    }, TIME_TO_RENDER_TREE)
   }, [isSegmentLoaded, pathname])
   return null
 }
